==> scripts/old-analyses/skagerak/03_get_skagerak_trawling.R <==
############################################################################
#####               het_skagerak_trawling                ####
############################################################################
# 
# This scripts downloads the gridded effort data for Skagerak MPA
# from Google Big Query and then exports it as an rds file.
# 
############################################################################

## Set up
# Load packages
library(here)
library(connections)
library(bigrquery)
library(tidyverse)

# Create a connection to the blue halos project
# You will be prompted to authenticate using your authorized email
# After this, you'll see them appear in your connections pane, where you can explroe each project!
ocean_halos_v2 <- connection_open(
  bigquery(),
  project = "emlab-gcp",          # project you want to connect to
  dataset = "ocean_halos_v2",     # dataset we're working on
  billing = "emlab-gcp",          # Who's paying for this?
  use_legacy_sql = FALSE,         # Just don't
  allowLargeResults = TRUE        # Give me all the data
)

table_name <- "skagerak_trawling"

# Download the data (and export it)
data <- tbl(ocean_halos_v2, table_name) %>% 
  collect()

# Save the data
saveRDS(object = data,
        file = here("data", paste0(table_name, ".rds")))
